fix(geometry): conform seams across plane buckets — the T-junction root cause
consolidate_coplanar re-triangulates each coplanar plane bucket independently.
Where a bucket's boundary runs along a facet it shares with an abutting bucket, its
collinear simplify can drop a vertex the neighbour keeps: the shared edge is then one
long edge on one side against two short ones on the other. That is a T-junction, and
it renders as a hairline crack. dental_clinic #217 is the minimal case — a
12-triangle wall, watertight with no voids, comes out of one opening subtract with 6
unmatched half-edges.

THE DISCRIMINATION. A blanket 'protect shared vertices' rule fails: on the #098
reveal wall it drives unpaired edges from 42 to 324, because a faceted reveal shares
almost every vertex and blanket protection defeats the merge the simplify exists to
perform. Narrowing to non-parallel neighbours only reaches 298. What actually
separates the cases is the simplify's OWN judgment, read across buckets: a genuine
seam vertex is a hard corner in the abutting bucket, so it survives that bucket's
simplify and is recorded; an i_overlay phantom is near-collinear in every bucket that
touches it, is dropped everywhere, and so never becomes an insertion source. No
geometric proxy for 'is this a real corner' is needed.

Three phases over the existing bucket map: A unions/welds/simplifies and records
surviving ring vertices at 0.1 mm with the count of distinct buckets that kept each;
B inserts, per bucket, those kept by some OTHER bucket that lie on this bucket's
plane and strictly interior to a ring edge; C triangulates and emits in today's
order. Today's output is emitted FIRST, so an already-watertight host pays nothing.

TWO DECISIONS THAT ARE MEASURED, NOT CAUTIOUS:
- The accept bar is ALL-OR-NOTHING: take the conformed mesh only if it is fully
  watertight. Accepting partial improvements sends a half-conformed intermediate into
  the next sequential cut, which compounds — that is what produced the 324.
- The decision uses a 0.1 mm open-edge grid, not the 1 mm default. At 1 mm a chorded
  seam merges away and reads as closed, which is exactly what accepted the 324.

Split into consolidate/{conform,ring_ops}.rs, prism_cut/{closure_checks,vertex_dedup}.rs
and triangulation/{alt_oracle,ring_geom}.rs so every file SHRINKS under the
module-size ratchet rather than raising a budget: consolidate 990->825,
prism_cut 3251->2958, triangulation 798->763.

Measured over 116 fixtures / 1355 void-hosting elements:
  TOTAL unmatched edges  18252 -> 17792   (-460)
  torn void hosts          238 -> 199
  closed-solid defects      80 ->  41     (52 -> 28 boolean-caused, 28 -> 13 arrived torn)
  triangulator-sensitive   139 -> 136
  snap-collapsed hosts      55 ->  51
  #098 reveal wall      42 -> 22 unpaired edges (improved)
  dental_clinic #217     6 -> 0 unmatched edges (watertight)

Every metric strictly better; all five baselines re-pinned to the new values. 82
geometry targets + 35 processing targets green, clippy clean, mesh-determinism
manifest unchanged. One snapshot re-pinned (bath #780: 128->152 tris, 384->216 verts;
surface area, bbox and spike count identical).
